Issue #93: Stripe payment api moved from checkout.Session to PaymentIntent. Updated Stripe Payment Intents method for event registrations. Refactored views, templates, urls for memberships.
