+++ dds_registration/templates/dds_registration/billing/billing_membership_form.html.django
@@ -3,36 +3,32 @@
   @module billing_membership_form.html.django
   @changed 2024.04.01, 16:53
 -->
-
 {% extends "base-regular.html.django" %}
-
 {% load crispy_forms_tags %}
-
-{% block title %}Membership dues — {{ block.super }}{% endblock title %}
-
+{% block title %}
+  Membership dues — {{ block.super }}
+{% endblock title %}
 {% block content %}
-{{ block.super }}
-
-<fieldset>
-  <legend>
-    <h1 class="page-title primary-color">Membership dues</h1>
-  </legend>
-  {% if error_message %}<p><strong>{{ error_message }}</strong></p>{% endif %}
-</fieldset>
-
-{% include "assets/render-messages/render-messages.django" %}
-
-{# Place form... #}
-
-<form role="form" method="post">
-  {% csrf_token %}
-  {{ form|crispy }}
-
-  <div class="common-actions">
-    <input class="btn btn-primary" type="submit" value="Save" />
-    {% include "assets/back-button/back-button.django" %}
-  </div>
-
-</form>
-
+  {{ block.super }}
+  <fieldset>
+    <legend>
+      <h1 class="page-title primary-color">Membership dues</h1>
+    </legend>
+    {% if error_message %}
+      <p>
+        <strong>{{ error_message }}</strong>
+      </p>
+    {% endif %}
+  </fieldset>
+  {% include "assets/render-messages/render-messages.django" %}
+  {# Place form... #}
+  <form role="form" method="post">
+    {% csrf_token %}
+    {{ form|crispy }}
+    <div class="common-actions">
+      {# TODO: Issue #71: Change the submit button texts dynamically depending on the payment method control: 'Save' for 'INVOICE' and 'Pay now' for 'STRIPE'. #}
+      <input class="btn btn-primary" type="submit" value="Pay now / Save" />
+      {% include "assets/back-button/back-button.django" %}
+    </div>
+  </form>
 {% endblock content %}
