+++ dds_registration/urls/billing_membership_stripe.py
@@ -19,14 +19,8 @@ urlpatterns = [
         name="billing_membership_stripe_payment_proceed",
     ),
     path(
-        # Create stripe session (Link with membership type, currency and amount)
-        "billing/membership/<str:membership_type>/payment/stripe/create_checkout_session/<str:currency>/<float:amount>",
-        billing_membership_stripe_views.billing_membership_payment_stripe_create_checkout_session,
-        name="billing_membership_payment_stripe_create_checkout_session",
-    ),
-    path(
         # Stripe payment success
-        "billing/membership/<str:membership_type>/payment/stripe/success/<str:session_id>",
+        "billing/membership/payment/stripe/success",
         billing_membership_stripe_views.billing_membership_stripe_payment_success,
         name="billing_membership_stripe_payment_success",
     ),
