+++ dds_registration/urls/membership.py
@@ -13,12 +13,6 @@ urlpatterns = [
         membership_views.membership_start,
         name="membership_start",
     ),
-    # path(
-    #     # UNUSED?
-    #     "membership/proceed/<str:membership_type>",
-    #     membership_views.membership_proceed,
-    #     name="membership_proceed",
-    # ),
     path(
         "membership/proceed/success",
         membership_views.membership_proceed_success,
