+++ src/assets/stripe-init/billing_membership_stripe_payment_proceed.ts
@@ -1,41 +1,9 @@
 /**
  * @module billing_membership_stripe_payment_proceed.ts
- * @changed 2024.04.03, 16:18
+ * @changed 2024.04.04, 01:15
  */
 
-export function billing_membership_stripe_payment_proceed(params: TCreateCheckoutSessionParams) {
-  const {
-    STRIPE_PUBLISHABLE_KEY,
-    success_url,
-  } = params;
+import { startStripeElementsForm } from './stripe_payment_intents_support'
 
-  console.log('[billing_membership_stripe_payment_proceed]', {
-    STRIPE_PUBLISHABLE_KEY,
-    success_url,
-  });
-
-  // Initialize Stripe.js
-  const stripe = window.Stripe(STRIPE_PUBLISHABLE_KEY);
-
-  initialize();
-
-  // Fetch Checkout Session and retrieve the client secret
-  async function initialize() {
-    const fetchClientSecret = async () => {
-      const response = await fetch(success_url, {
-        method: 'POST',
-      });
-      const { clientSecret } = await response.json();
-      return clientSecret;
-    };
-
-    // Initialize Checkout
-    const checkout = await stripe.initEmbeddedCheckout({
-      fetchClientSecret,
-    });
-
-    // Mount Checkout
-    checkout.mount('#checkout');
-  }
-
-}
+export const billing_membership_stripe_payment_proceed = startStripeElementsForm;
+// NOTE: It's possible to use `startStripeElementsForm` directly here.
